requirements.txt
requests
pymupdf
sentence-transformers
faiss-cpu
fastapi
uvicorn
openai
